Lab4/vjezba4/src/main/java/com/oss/vjezba4/controllers/ClientController.java
package com.oss.vjezba4.controllers;


import com.oss.vjezba4.dto.ClientPayload;
import com.oss.vjezba4.models.Client;
import com.oss.vjezba4.services.ClientService;
import com.oss.vjezba4.utils.ResponseMessage;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping(path = "/api/client")
public class ClientController {
    private final ClientService clientService;

    ClientController(ClientService clientService) {
        this.clientService = clientService;
    }

    @PostMapping
    public ResponseEntity<?> createClient(@RequestBody ClientPayload payload) {
        var client = clientService.createClient(payload);
        if (client != null) {
            return new ResponseEntity<>(client, HttpStatus.OK);
        } else {
            return new ResponseEntity<>(new ResponseMessage("Can not create client"), HttpStatus.BAD_REQUEST);
        }
    }

    @GetMapping("/{id}")
    public ResponseEntity<?> getClientById(@PathVariable Long id) {
        var client = clientService.getClientById(id);
        if (client != null) {
            return new ResponseEntity<>(client, HttpStatus.OK);
        } else {
            return new ResponseEntity<>(new ResponseMessage("Client with id: " + id + " does not exist"), HttpStatus.NOT_FOUND);
        }
    }

    @GetMapping
    public List<Client> getAllClients() {
        return clientService.getAllClients();
    }
}
